requirements.txt
torch==1.7.1
pandas==1.3
biopython
scikit-learn==0.24.2
matplotlib
lxml
requests


-r requirements.txt
biovec
fair-esm
-e git+https://github.com/amyxlu/CPCProt.git#egg=CPCProt